MB-2009-09-25-0: Fix.      : NPE when a JMX connection was created to a host that is not longer resolvable/accessible
git-svn-id: https://svn.java.net/svn/visualvm~svn/branches/sgemacc@2331 ba0df5e5-9053-6d9f-941f-f112716aa91a

+++ visualvm/jmx/src/com/sun/tools/visualvm/jmx/application/JmxApplicationProvider.java
@@ -291,7 +291,7 @@ public class JmxApplicationProvider {
         // NOTE: this is not absolutely failsafe, if resolving the JMX application
         // took a long time and its host has been added by the user/plugin, it may
         // be removed by this call. Hopefully just a hypothetical case...
-        if (host != null && !hosts.contains(host))
+        if (host != null && !hosts.contains(host) && host.getOwner() != null)
             host.getOwner().getRepository().removeDataSource(host);
     }
